+++ src/components/Link.tsx
@@ -6,11 +6,12 @@ import { useStore } from '../store/storeContext';
 import { linkIsPublic, linkTargetUrl } from '../utils/links';
 import Icon from './Icon';
 
-const Content = styled.div<{ center?: boolean }>`
+const Content = styled.div<{ center?: boolean; linkStyle: 'normal' | 'alert' }>`
   display: flex;
   justify-content: ${p => (p.center ? 'center' : 'space-between')};
   align-items: center;
   width: 100%;
+  color: ${p => (p.linkStyle === 'alert' ? p.theme.color.accent : undefined)};
 
   .link__label-container {
     min-height: 28px;
@@ -38,8 +39,20 @@ const Content = styled.div<{ center?: boolean }>`
   }
 `;
 
+const LinkButton = styled.button`
+  margin: 0 !important;
+  padding: 0 !important;
+`;
+
+// Extend basic link model with additional "button" type
+export interface LinkItem extends Omit<LinkIn, 'type'> {
+  type: 'external' | 'internal' | 'page' | 'test' | 'exercise' | 'button';
+  onClick?: () => void;
+  style?: 'normal' | 'alert';
+}
+
 interface Props {
-  link: LinkIn;
+  link: LinkItem;
   label?: string | null;
   showArrow?: boolean;
   center?: boolean;
@@ -58,13 +71,15 @@ const Link: React.FC<Props> = ({
 
   if (!link) return null;
 
+  const isButton = link.type === 'button' && !!link.onClick;
+
   const isExternal = link.type === 'external';
-  const url = linkTargetUrl(link);
-  const isPublic = linkIsPublic(link);
+  const url = !isButton && linkTargetUrl(link as LinkIn);
+  const isPublic = !isButton && linkIsPublic(link as LinkIn);
   const showLock = !isPublic && !isLoggedIn;
 
   const content = (
-    <Content center={center}>
+    <Content center={center} linkStyle={link.style ?? 'normal'}>
       {label && (
         <div className="link__label-container">
           {showLock && <Icon type="Lock" />}
@@ -86,6 +101,14 @@ const Link: React.FC<Props> = ({
       </div>
     </Content>
   );
+
+  if (isButton) {
+    return (
+      <LinkButton role="link" onClick={link.onClick}>
+        <span className="link">{content}</span>
+      </LinkButton>
+    );
+  }
 
   if (!url) return null;